+++ app/src/main/java/com/example/idus/ui/model/WeatherInformation.kt
@@ -2,4 +2,4 @@ package com.example.idus.ui.model
 
 data class WeatherInformation(val location: String, val woeid: Int, val today: WeatherItem, val tomorrow: WeatherItem)
 
-data class WeatherItem(val iconUrl: String, val temperature: Double, val humidity: Int, val weather: String)
+data class WeatherItem(val iconUrl: String, val temperature: Int, val humidity: Int, val weather: String)
